requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
sqlalchemy
streamlit
requests
streamlit-autorefresh
